requirements.txt
Django==1.11
djangorestframework==3.8.2
Pillow==5.2.0
pytz==2018.5
